Separate binary and library code.
Move code outside of the `main.rs` to `lib/mod.rs` file to make
`JDServer` accessible through other enviornements.
